Add check for nonexistant directories
This MR fixes an issue that occurs when resetting the state when one or more of the expected directories is not present (e.g., the "visibilities" directory is created only when visibilities are saved).

See merge request RD/rapthor!433

+++ rapthor/modifystate.py
@@ -128,11 +128,12 @@ def run(parset_file):
                 "images",
                 "visibilities",
             ):
-                dcmp = filecmp.dircmp(
-                    os.path.join(parset['dir_working'], 'pipelines'),
-                    os.path.join(parset['dir_working'], dirname)
-                )
-                for path in (os.path.join(dcmp.right, item) for item in dcmp.right_only):
-                    shutil.rmtree(path, ignore_errors=True)
+                if os.path.exists(os.path.join(parset['dir_working'], dirname)):
+                    dcmp = filecmp.dircmp(
+                        os.path.join(parset['dir_working'], 'pipelines'),
+                        os.path.join(parset['dir_working'], dirname)
+                    )
+                    for path in (os.path.join(dcmp.right, item) for item in dcmp.right_only):
+                        shutil.rmtree(path, ignore_errors=True)
 
             print('Reset complete.')
